Allow different broker implementations to re-use the same test cases by abstracting broker specific functions.
